# api/openapi.yaml
openapi: 3.0.0
info:
  title: Pahlawan Pangan Public API
  description: API Platform untuk Penyelamatan Makanan Surplus Skala Nasional Indonesia.
  version: 2.0.0
  contact:
    name: Engineering Team Pahlawan Pangan
    email: engineering@pahlawanpangan.org
servers:
  - url: https://api.pahlawanpangan.org/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development
paths:
  /marketplace:
    get:
      summary: Cari Makanan Murah (B2C)
      description: Mendapatkan daftar makanan surplus di sekitar lokasi pengguna dengan harga diskon dinamis.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Daftar makanan ditemukan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurplusItem'
  /surplus:
    post:
      summary: Posting Makanan Surplus (Provider Only)
      description: Restoran atau Hotel memposting kelebihan makanan mereka.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostSurplusRequest'
      responses:
        '201':
          description: Berhasil diposting
...
  /merchant/verify-pickup:
    post:
      summary: Verifikasi QR Code Pickup (Merchant)
      description: Memvalidasi kode OTP/QR yang dibawa pembeli untuk mengonfirmasi serah terima makanan.
      responses:
        '200':
          description: Verifikasi Berhasil dan Status Pesanan Selesai
        '404':
          description: Kode Tidak Valid atau Sudah Digunakan
components:
  schemas:
    SurplusItem:
      type: object
      properties:
        id:
          type: string
        food_type:
          type: string
        original_price:
          type: number
        current_price:
          type: number
        expires_in:
          type: number
          description: Menit menuju kedaluwarsa
    PostSurplusRequest:
      type: object
      properties:
        provider_id:
          type: string
        food_type:
          type: string
        quantity_kgs:
          type: number
        original_price:
          type: number
        expiry_time:
          type: string
          format: date-time