# internal/api/handler.go
package api

import (
	"context"
	"database/sql"
	"fmt"
	"net/http"
	"time"

	"github.com/albnnaardy11/pahlawan-pangan/internal/matching"
	"github.com/albnnaardy11/pahlawan-pangan/internal/outbox"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	"github.com/google/uuid"
	"github.com/segmentio/encoding/json"
	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
	"go.opentelemetry.io/otel"
	"go.opentelemetry.io/otel/attribute"
)

var tracer = otel.Tracer("api-handler")

type Handler struct {
	db            *sql.DB
	matchEngine   *matching.MatchingEngine
	outboxService *outbox.OutboxService
}

func NewHandler(db *sql.DB, engine *matching.MatchingEngine, outbox *outbox.OutboxService) *Handler {
	return &Handler{
		db:            db,
		matchEngine:   engine,
		outboxService: outbox,
	}
}

func (h *Handler) Routes() http.Handler {
	r := chi.NewRouter()

	// Middleware
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)
	r.Use(middleware.Timeout(30 * time.Second))

	// CORS configuration for Frontend Devs
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{"https://*", "http://*"},
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token"},
		ExposedHeaders:   []string{"Link"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	// Health checks
	r.Get("/health/live", h.LivenessCheck)
	r.Get("/health/ready", h.ReadinessCheck)

	// API routes with OpenTelemetry instrumentation
	r.Route("/api/v1", func(r chi.Router) {
		r.Use(otelhttp.NewMiddleware("api"))

		// Surplus endpoints
		r.Post("/surplus", h.PostSurplus)
		r.Get("/surplus/{id}", h.GetSurplus)
		r.Post("/surplus/{id}/claim", h.ClaimSurplus)
		r.Get("/marketplace", h.BrowseSurplus)

		// Social & Pahlawan-AI Unicorn Features
		r.Get("/feed", h.GetSocialFeed)
		r.Get("/analytics/predict", h.GetWastePrediction)

		// Next-Gen Super-App Features
		r.Post("/express/request", h.RequestExpress)
		r.Post("/integration/pos/sync", h.SyncPOS)
		r.Get("/sustainability/carbon-report", h.GetCarbonReport)
		r.Post("/community/group-buy/join", h.JoinGroupBuy)
		r.Get("/community/drop-points", h.GetDropPoints)
		r.Get("/analytics/provider-roi", h.GetProviderROI)

		// Super-App Phase 2: Ratings, Vouchers & Chat
		r.Post("/ratings", h.AddRating)
		r.Post("/vouchers/apply", h.ApplyVoucher)
		r.Get("/marketplace/recommendations", h.GetRecommendations)
		r.Route("/chat", func(r chi.Router) {
			r.Post("/threads", h.OpenChat)
			r.Get("/threads", h.ListChats)
		})
		r.Put("/account/settings", h.UpdateSettings)

		// Merchant Dashboard (Provider-specific)
		r.Route("/merchant", func(r chi.Router) {
			r.Get("/claims", h.GetProviderClaims) // List all claims for this provider
			r.Post("/verify-pickup", h.VerifyPickupCode) // Scan/Verify QR code
			r.Get("/analytics", h.GetProviderROI) // Integrated ROI analytics
		})

		// NGO endpoints
		r.Get("/ngos/nearby", h.GetNearbyNGOs)
	})

	return r
}

type PostSurplusRequest struct {
	ProviderID  string    `json:"provider_id"`
	Lat         float64   `json:"lat"`
	Lon         float64   `json:"lon"`
	QuantityKgs float64   `json:"quantity_kgs"`
	FoodType    string    `json:"food_type"`
	ExpiryTime  time.Time `json:"expiry_time"`
}

func (h *Handler) PostSurplus(w http.ResponseWriter, r *http.Request) {
	ctx, span := tracer.Start(r.Context(), "PostSurplus")
	defer span.End()

	var req PostSurplusRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	// Validate
	if req.QuantityKgs <= 0 || req.ExpiryTime.Before(time.Now()) {
		http.Error(w, "invalid request", http.StatusBadRequest)
		return
	}

	surplusID := uuid.New().String()

	// Start transaction
	tx, err := h.db.BeginTx(ctx, nil)
	if err != nil {
		http.Error(w, "database error", http.StatusInternalServerError)
		return
	}
	defer tx.Rollback()

	// Insert surplus
	_, err = tx.ExecContext(ctx, `
		INSERT INTO surplus (id, provider_id, location, quantity_kgs, food_type, expiry_time, status, geo_region_id, created_at)
		SELECT $1, $2, ST_SetSRID(ST_MakePoint($3, $4), 4326), $5, $6, $7, 'available',
		       (SELECT id FROM geo_regions WHERE ST_Contains(geometry, ST_SetSRID(ST_MakePoint($3, $4), 4326)) LIMIT 1),
		       NOW()
	`, surplusID, req.ProviderID, req.Lon, req.Lat, req.QuantityKgs, req.FoodType, req.ExpiryTime)

	if err != nil {
		span.RecordError(err)
		http.Error(w, "failed to create surplus", http.StatusInternalServerError)
		return
	}

	// Create outbox event
	payload, _ := json.Marshal(map[string]interface{}{
		"surplus_id":   surplusID,
		"provider_id":  req.ProviderID,
		"lat":          req.Lat,
		"lon":          req.Lon,
		"quantity_kgs": req.QuantityKgs,
		"expiry_time":  req.ExpiryTime,
	})

	event := outbox.OutboxEvent{
		ID:          uuid.New().String(),
		AggregateID: surplusID,
		EventType:   outbox.SurplusPosted,
		Payload:     payload,
	}

	err = h.outboxService.PublishWithTransaction(ctx, tx, event)
	if err != nil {
		span.RecordError(err)
		http.Error(w, "failed to publish event", http.StatusInternalServerError)
		return
	}

	// Commit transaction
	if err := tx.Commit(); err != nil {
		http.Error(w, "failed to commit", http.StatusInternalServerError)
		return
	}

	span.SetAttributes(attribute.String("surplus_id", surplusID))

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"surplus_id": surplusID,
		"status":     "posted",
	})
}

type ClaimSurplusRequest struct {
	NGOID             string  `json:"ngo_id"`
	FulfillmentMethod string  `json:"fulfillment_method"` // 'courier' or 'self_pickup'
	UserLat           float64 `json:"user_lat"`
	UserLon           float64 `json:"user_lon"`
}

func (h *Handler) ClaimSurplus(w http.ResponseWriter, r *http.Request) {
	ctx, span := tracer.Start(r.Context(), "ClaimSurplus")
	defer span.End()

	surplusID := chi.URLParam(r, "id")

	// Weakness Fix: Check Liability Waiver
	waiver := r.Header.Get("X-Liability-Waiver-Accepted")
	if waiver != "true" {
		http.Error(w, "Legal: You must accept the Food Safety Liability Waiver", http.StatusForbidden)
		return
	}

	var req ClaimSurplusRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	// Complex Orchestration: Logistics vs Self-Pickup
	// Note: In real app, fetch storeLat/Lon from DB using surplusID
	storeLat, storeLon := -6.2, 106.8 // Mock coords for Jakarta
	
	fOption := matching.FulfillmentOption(req.FulfillmentMethod)
	fStatus, err := matching.OrchestrateFulfillment(fOption, req.UserLat, req.UserLon, storeLat, storeLon)
	if err != nil {
		http.Error(w, fmt.Sprintf("Fulfillment Error: %v", err), http.StatusUnprocessableEntity)
		return
	}

	// Update DB (Optimistic Locking)
	result, err := h.db.ExecContext(ctx, `
		UPDATE surplus 
		SET status = 'claimed', 
		    claimed_by_ngo_id = $1, 
		    claimed_at = NOW(),
		    version = version + 1
		WHERE id = $2 
		  AND status = 'available'
		  AND expiry_time > NOW()
	`, req.NGOID, surplusID)

	if err != nil {
		http.Error(w, "database error", http.StatusInternalServerError)
		return
	}

	rows, _ := result.RowsAffected()
	if rows == 0 {
		http.Error(w, "surplus already claimed or expired", http.StatusConflict)
		return
	}

	// Create Delivery/Pickup Record
	_, err = h.db.ExecContext(ctx, `
		INSERT INTO deliveries (surplus_id, fulfillment_method, pickup_verification_code, external_tracking_id, status)
		VALUES ($1, $2, $3, $4, 'assigned')
	`, surplusID, fStatus.Method, fStatus.VerificationCode, fStatus.TrackingID)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"status":      "claimed",
		"fulfillment": fStatus,
	})
}


// BrowseSurplus allows general citizens to find cheap food (B2C Unicorn feature)
func (h *Handler) BrowseSurplus(w http.ResponseWriter, r *http.Request) {
	ctx, span := tracer.Start(r.Context(), "BrowseSurplus")
	defer span.End()

	_ = r.URL.Query().Get("lat")
	_ = r.URL.Query().Get("lon")

	// In production, use PostGIS:
	// SELECT * FROM surplus WHERE ST_DWithin(location, ST_MakePoint($1, $2), 5000)
	// AND status = 'available'

	rows, err := h.db.QueryContext(ctx, `
		SELECT id, provider_id, original_price, discount_price, food_type, expiry_time, temperature_category 
		FROM surplus 
		WHERE status = 'available' AND expiry_time > NOW()
		LIMIT 20
	`)
	if err != nil {
		http.Error(w, "Failed to fetch marketplace", http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	var results []map[string]interface{}
	for rows.Next() {
		var item struct {
			ID            string
			ProviderID    string
			OriginalPrice float64
			DiscountPrice float64
			FoodType      string
			ExpiryTime    time.Time
		}
		rows.Scan(&item.ID, &item.ProviderID, &item.OriginalPrice, &item.DiscountPrice, &item.FoodType, &item.ExpiryTime)

		results = append(results, map[string]interface{}{
			"id":             item.ID,
			"food_type":      item.FoodType,
			"original_price": item.OriginalPrice,
			"current_price":  item.DiscountPrice,
			"expires_in":     time.Until(item.ExpiryTime).Minutes(),
		})
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(results)
}

// GetSocialFeed returns a Strava-style feed of food rescues (Shared Social Proof)
func (h *Handler) GetSocialFeed(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "GetSocialFeed")
	defer span.End()

	// Logic: Fetch recent successful rescues with photos
	results := []map[string]interface{}{
		{
			"user":      "Budi Penyelamat",
			"action":    "Rescued 5kg Bakery items from 'Roti Enak Jaksel'",
			"impact":    "Saved 12kg of CO2",
			"cheers":    42,
			"media_url": "https://cdn.Pahlawan Pangan.org/rescuers/budi_1.jpg",
		},
		{
			"user":      "Santi Zero-Waste",
			"action":    "Donated 20 boxes of meals to 'Panti Asuhan Kasih'",
			"impact":    "Fed 40 children",
			"cheers":    156,
			"media_url": "https://cdn.Pahlawan Pangan.org/rescuers/santi_2.jpg",
		},
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(results)
}

// GetWastePrediction uses AI to tell restaurants if they will have leftovers today
func (h *Handler) GetWastePrediction(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "GetWastePrediction")
	defer span.End()

	providerID := r.URL.Query().Get("provider_id")
	if providerID == "" {
		http.Error(w, "provider_id is required", http.StatusBadRequest)
		return
	}

	// In real setup, h.aiEngine.PredictWaste(ctx, providerID)
	prediction := map[string]interface{}{
		"provider_id":        providerID,
		"predicted_loss_kgs": 12.5,
		"confidence":         0.89,
		"reason":             "Rainy weather predicted in South Jakarta (decreases walk-in customers)",
		"recommendation":     "Active 'Flash Ludes' at 19:00 with 70% discount",
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(prediction)
}

// Next-Gen Feature Handlers

func (h *Handler) RequestExpress(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "RequestExpress")
	defer span.End()

	// Weakness Fix: Check Cold Chain Requirement
	temp := r.URL.Query().Get("temp")
	requiresColdChain := false
	if temp == "chilled" || temp == "frozen" || temp == "hot" {
		requiresColdChain = true
	}

	// Simulating logistics request
	res := map[string]interface{}{
		"status":              "courier_searching",
		"service":             "Pahlawan-Express",
		"requires_cold_chain": requiresColdChain,
		"impact_pts":          50,
		"est_pickup":          "12 minutes",
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) SyncPOS(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "SyncPOS")
	defer span.End()

	// Simulating POS automation
	res := map[string]interface{}{
		"provider":    "Bakery Center",
		"items_found": 12,
		"auto_posted": true,
		"sync_status": "success",
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) GetCarbonReport(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "GetCarbonReport")
	defer span.End()

	res := map[string]interface{}{
		"co2_saved_kg":  450.5,
		"esg_tokens":    45,
		"certification": "Zero Waste Gold",
		"period":        "Monthly",
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) JoinGroupBuy(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "JoinGroupBuy")
	defer span.End()

	res := map[string]interface{}{
		"group_id": "RT05-RW02-SUDIRMAN",
		"members":  12,
		"total_kg": 15.0,
		"goal_kg":  20.0,
		"status":   "forming",
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) GetDropPoints(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "GetDropPoints")
	defer span.End()

	res := []map[string]interface{}{
		{
			"id":       "DP-001",
			"name":     "Pos Satpam Cluster Sakura",
			"address":  "Jl. Sudirman No. 1",
			"distance": "150m",
		},
		{
			"id":       "DP-002",
			"name":     "Rumah Ketua RT 05",
			"address":  "Gg. Pahlawan 3",
			"distance": "420m",
		},
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) GetProviderROI(w http.ResponseWriter, r *http.Request) {
	_, span := tracer.Start(r.Context(), "GetProviderROI")
	defer span.End()

	res := map[string]interface{}{
		"provider_id":     "P-777",
		"total_idr_saved": 12500000.0,
		"waste_saved_kg":  450.0,
		"co2_offset_kg":   1125.0,
		"eco_hero_status": "Guardian of the Green",
		"impact_ranking":  "Top 5% in Jakarta",
	}
	json.NewEncoder(w).Encode(res)
}

func (h *Handler) GetSurplus(w http.ResponseWriter, r *http.Request) {
	// Implementation omitted for brevity
	w.WriteHeader(http.StatusOK)
}

func (h *Handler) GetNearbyNGOs(w http.ResponseWriter, r *http.Request) {
	// Implementation omitted for brevity
	w.WriteHeader(http.StatusOK)
}

func (h *Handler) LivenessCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusOK)
	w.Write([]byte("OK"))
}

func (h *Handler) ReadinessCheck(w http.ResponseWriter, r *http.Request) {
	// Check database connection
	ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
	defer cancel()

	if err := h.db.PingContext(ctx); err != nil {
		http.Error(w, "database not ready", http.StatusServiceUnavailable)
		return
	}

	w.WriteHeader(http.StatusOK)
	w.Write([]byte("READY"))
}

// --- Super-App Phase 2 Handlers (Standard Gojek/Tokopedia) ---

func (h *Handler) AddRating(w http.ResponseWriter, r *http.Request) {
	// Logic for adding ratings to providers/couriers
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "rating_submitted"})
}

func (h *Handler) ApplyVoucher(w http.ResponseWriter, r *http.Request) {
	// Logic for validating pahlawan vouchers
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{"status": "valid", "discount_idr": 15000})
}

func (h *Handler) GetRecommendations(w http.ResponseWriter, r *http.Request) {
	// Logic for weighted ranking (RecEngine)
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{"note": "showing top-ranked food rescue items"})
}

func (h *Handler) OpenChat(w http.ResponseWriter, r *http.Request) {
	// Meta-logic for starting chat threads with driver/hotel
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"thread_id": uuid.New().String()})
}

func (h *Handler) ListChats(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode([]string{"active_thread_1", "active_thread_2"})
}

func (h *Handler) UpdateSettings(w http.ResponseWriter, r *http.Request) {
	// Logic for user/provider account settings
	w.WriteHeader(http.StatusOK)
}

// --- Merchant Dashboard Handlers ---

func (h *Handler) GetProviderClaims(w http.ResponseWriter, r *http.Request) {
	// Logic to list all active/completed claims for the logged-in provider
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"active_claims":    5,
		"completed_today": 12,
	})
}

func (h *Handler) VerifyPickupCode(w http.ResponseWriter, r *http.Request) {
	ctx, span := tracer.Start(r.Context(), "VerifyPickupCode")
	defer span.End()

	var req struct {
		VerificationCode string `json:"verification_code"`
		ProviderID       string `json:"provider_id"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	// Complex Logic: Verify code in DB for specific provider
	result, err := h.db.ExecContext(ctx, `
		UPDATE deliveries 
		SET is_verified_pickup = true, status = 'delivered', updated_at = NOW()
		FROM surplus
		WHERE deliveries.surplus_id = surplus.id
		  AND surplus.provider_id = $1
		  AND deliveries.pickup_verification_code = $2
		  AND deliveries.is_verified_pickup = false
	`, req.ProviderID, req.VerificationCode)

	if err != nil {
		http.Error(w, "Verification failed", http.StatusInternalServerError)
		return
	}

	rows, _ := result.RowsAffected()
	if rows == 0 {
		http.Error(w, "Invalid or already used verification code", http.StatusNotFound)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status":  "verified",
		"message": "Pickup successful. Inventory updated.",
	})
}